requirements.txt
aiohttp
aiohttp-retry
aiosignal
async-timeout
attrs
blinker
certifi
charset-normalizer
click
colorama
Flask
frozenlist
idna
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
multidict
openai
PyJWT
python-dotenv
pytz
requests
tqdm
twilio
urllib3
Werkzeug
yarl
zipp
